refactor TroubleshooterMessageDialog to extend ErrorMessageDialog
